добавлена проверка наличия уведомления по ID в GET, PUT, DELETE; добавлен коммент TODO для переделки метода полчучения email в MailServiceImpl

+++ web/src/main/java/com/vet24/web/controllers/notification/AdminNotificationController.java
@@ -14,8 +14,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.webjars.NotFoundException;
 
-import javax.validation.Valid;
 import java.util.List;
 
 @RestController
@@ -53,8 +53,13 @@ public class AdminNotificationController {
     @GetMapping(value = {"/{id}"})
     public ResponseEntity<NotificationDto> getNotificationById(
             @PathVariable("id") Long notificationId) {
-        Notification notification = notificationService.getByKey(notificationId);
-        return new ResponseEntity<>(notificationMapper.toDto(notification), HttpStatus.OK);
+        if(notificationService.isExistByKey(notificationId)) {
+            Notification notification = notificationService.getByKey(notificationId);
+            return new ResponseEntity<>(notificationMapper.toDto(notification), HttpStatus.OK);
+        } else {
+            log.info("Notification with id {} not found", notificationId);
+            throw new NotFoundException("Notification not found");
+        }
     }
 
     @Operation(summary = "Update the notification by ID")
@@ -65,12 +70,16 @@ public class AdminNotificationController {
     @PutMapping(value = {"/{id}"})
     public ResponseEntity<NotificationDto> updateNotificationById(
             @PathVariable("id") Long notificationId,
-            @RequestBody(required = false)
-            @Valid NotificationDto notificationDto) {
-        Notification notification = notificationService.getByKey(notificationId);
-        notificationMapper.updateEntity(notificationDto, notification);
-        notificationService.update(notification);
-        return new ResponseEntity<>(notificationMapper.toDto(notification), HttpStatus.OK);
+            @RequestBody NotificationDto notificationDto) {
+        if(notificationService.isExistByKey(notificationId)) {
+            Notification notification = notificationService.getByKey(notificationId);
+            notificationMapper.updateEntity(notificationDto, notification);
+            notificationService.update(notification);
+            return new ResponseEntity<>(notificationMapper.toDto(notification), HttpStatus.OK);
+        } else {
+            log.info("Notification with id {} not found", notificationId);
+            throw new NotFoundException("Notification not found");
+        }
     }
 
 
@@ -81,9 +90,7 @@ public class AdminNotificationController {
             @ApiResponse(responseCode = "404", description = "Notification not found"),
     })
     @PostMapping
-    public ResponseEntity<NotificationDto> createNotification(
-            @RequestBody(required = false)
-            @Valid NotificationDto notificationDto) {
+    public ResponseEntity<NotificationDto> createNotification(@RequestBody NotificationDto notificationDto) {
         Notification notification = notificationMapper.toEntity(notificationDto);
         notificationService.persist(notification);
         return ResponseEntity.status(200).body(notificationMapper.toDto(notification));
@@ -96,10 +103,15 @@ public class AdminNotificationController {
             @ApiResponse(responseCode = "404", description = "Notification not founded")
     })
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteNotification(@PathVariable("id") Long id) {
-        Notification notification = notificationService.getByKey(id);
-        notificationService.delete(notification);
-        return ResponseEntity.ok().build();
+    public ResponseEntity<Void> deleteNotification(@PathVariable("id") Long notificationId) {
+        if(notificationService.isExistByKey(notificationId)) {
+            Notification notification = notificationService.getByKey(notificationId);
+            notificationService.delete(notification);
+            return ResponseEntity.ok().build();
+        } else {
+            log.info("Notification with id {} not found", notificationId);
+            throw new NotFoundException("Notification not found");
+        }
     }
 
 }
